+++ Code/Libraries/Foundation/SpatialPartition/Sap.inl
@@ -315,8 +315,7 @@ void Sap<ClientDataType>::Validate()
       ErrorIf(endPoint.GetIndex() != index, "Endpoint index does not point to its index.");
       ErrorIf(mBoxes[endPoint.GetIndex() / 6].mObj == nullptr, "Endpoint points to a deleted box");
 
-      // validate that the sentinels are in the right spot and flagged min/max
-      // correctly
+      // validate that the sentinels are in the right spot and flagged min/max correctly
       if (i == 0)
       {
         // check the max
@@ -418,7 +417,7 @@ uint Sap<ClientDataType>::GetNewBoxIndex()
   {
     // add a new box and calculate its index
     mBoxes.PushBack();
-    index = mBoxes.Size() - 1;
+    index = (uint)mBoxes.Size() - 1;
 
     // add 6 indices for the endpoints
     for (uint i = 0; i < 6; ++i)
@@ -549,9 +548,9 @@ uint Sap<ClientDataType>::BatchSort(EndPointArray& newEndpoints)
 
   EndPointArray& axis = mAxes[Axis];
 
-  uint oldPos = axis.Size() - 1;
+  uint oldPos = (uint)axis.Size() - 1;
   axis.Resize(axis.Size() + newEndpoints.Size());
-  uint insertIndex = axis.Size() - 1;
+  uint insertIndex = (uint)axis.Size() - 1;
 
   // loop from the end to the front while we still have new endpoints to Insert
   Array<EndPointType>::range newRange = newEndpoints.All();
@@ -591,7 +590,7 @@ void Sap<ClientDataType>::BatchPairAdd(uint startIndex)
 
   EndPointArray& axis = mAxes[Axis];
   // make sure to avoid the sentinel node
-  uint lastEndPoint = axis.Size() - 2;
+  uint lastEndPoint = (uint)axis.Size() - 2;
   for (uint i = lastEndPoint; i >= startIndex; --i)
   {
     EndPointType& endPoint = axis[i];
@@ -642,7 +641,7 @@ void Sap<ClientDataType>::BatchPairRemove()
   BoxSet normalBoxes, removeBoxes;
 
   // make sure to ignore the sentinel
-  uint size = axis.Size() - 1;
+  uint size = (uint)axis.Size() - 1;
   for (uint i = 1; i < size; ++i)
   {
     EndPointType& endpoint = axis[i];
@@ -685,7 +684,7 @@ void Sap<ClientDataType>::BatchEndPointRemove()
 {
   EndPointArray& axis = mAxes[Axis];
   uint startIndex = 1;
-  uint endIndex = axis.Size() - 1;
+  uint endIndex = (uint)axis.Size() - 1;
   uint insertPosition = startIndex;
 
   // loop from beginning to end, ignoring the sentinels
@@ -731,7 +730,7 @@ template <uint Axis>
 void Sap<ClientDataType>::InsertEndpointOnAxis(const EndPointType& endpoint, bool finalAxis)
 {
   // get the index of the old last endpoint
-  int i = mAxes[Axis].Size() - 1;
+  int i = (int)mAxes[Axis].Size() - 1;
   // put the endpoint on the back of this axis
   mAxes[Axis].PushBack(endpoint);
 
@@ -746,8 +745,7 @@ void Sap<ClientDataType>::InsertEndpointOnAxis(const EndPointType& endpoint, boo
     --i;
   }
 
-  // set its index the index of the endpoint now that it is in its final
-  // position
+  // set its index the index of the endpoint now that it is in its final position
   mIndices[endpoint.GetIndex()] = i + 1;
 }
